backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
pydantic>=1.10
redis>=4.6.0
pytest>=7.0
httpx>=0.24
python-multipart>=0.0.6
bcrypt>=4.0.0
PyJWT>=2.8.0
